fix #28: Create spreading between cells. fix #29: add grid copy funtionallity

+++ cpu_implementation/cpu_cell.h
@@ -5,5 +5,5 @@ typedef struct
 {
     cp_sdl_cell graphics;
     int32_t fuel;
-    int32_t temprature;
+    float temprature;
 } cpu_cell;
